Fixed changes mentioned in pull request.
Not sure if that is how you wanted the comments of the params for the two endpoints?
I put required and optional params. Let us know if you would like us to add that for
all endpoints.

Note: I tested adding null or empty strings to fields in the mongodb shell we were
showed in lab. Seemed to store those values no problem. But I am not sure what will
be displayed on the front end. If these need to hold something different, it shouldnt
be hard to change.

Let us know if there is anything else.